refactor(arguments): extract resolvers to their own function
Because resolvers now are their own functions, all the ___Channel types
now depend on the base resolveChannel() resolver. This means each
one of them take a string as the input. Therefore, extended arguments
don't make any sense anymore as their sole purpose was to inherit the
"resolving" of the parent argument.