fix(Listener): deprecate event generic for Listener class
TS 4.5 no longer lets us compile @sapphire/framework when we have this generic
the way we used to have it. Seeing as we have plans to fully remove the generic
anyway in the next major release, we'll partially remove it for now by not reading
it anywhere in the code, and having it extend `any`.