fix(Args): ensure proper error types are always thrown
This changes the `result.unwrap()` calls to `result.unwrapRaw()` to
ensure that the erorr that gets thrown for the following methods:
- `args.pick(...)`
- `args.rest(...)`
- `args.repeat(...)`
- `args.peek(...)`
This makes it so that the error that ends up in `messageCommandError`
listener will have the proper error type. Furthermore, it can also
be handled properly when chaining `.catch()` calls to the `args.X` method.

resolves #528