feat: remove `CommandChatInputRegisterShortcut` alias interface, prefer using `ApplicationCommandRegistry.RegisterOptions` instead (#597)
The old interface exposed options that were not actually supported
by Sapphire, which was misleading. The whole interface could in fact
be removed, but that would be a breaking change, so for now we mark
it as deprecated and make it a type alias.

While normally removing interface options would be a breaking change too,
these particular options already did absolutely nothing and weren't
being read anywhere anymore. Anyone who still had this option and
will get a TypeScript build error can simply remove the options.